// packages/api-client/src/generated/models/GovernanceReasonRequest.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface GovernanceReasonRequest
 */
export interface GovernanceReasonRequest {
    /**
     *
     * @type {string}
     * @memberof GovernanceReasonRequest
     */
    reason: string;
}

/**
 * Check if a given object implements the GovernanceReasonRequest interface.
 */
export function instanceOfGovernanceReasonRequest(value: object): value is GovernanceReasonRequest {
    if (!('reason' in value) || value['reason'] === undefined) return false;
    return true;
}

export function GovernanceReasonRequestFromJSON(json: any): GovernanceReasonRequest {
    return GovernanceReasonRequestFromJSONTyped(json, false);
}

export function GovernanceReasonRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): GovernanceReasonRequest {
    if (json == null) {
        return json;
    }
    return {

        'reason': json['reason'],
    };
}

export function GovernanceReasonRequestToJSON(json: any): GovernanceReasonRequest {
    return GovernanceReasonRequestToJSONTyped(json, false);
}

export function GovernanceReasonRequestToJSONTyped(value?: GovernanceReasonRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'reason': value['reason'],
    };
}